MAINTENANCE: Raise the IOCTL version magic number
Because of the mismatch between upstream IgH EtherCAT master recently raising the IOCTL MAGIC version number to 31 and the patches from Gavin Lambert (4 years old now) raised the MAGIC number to 36, there is a possibility of issues with client software build against the older version 36. Thus, raise this number to 37 to avoid the possibility.